Adding examples with feedback icons
