+++ docs/app/pages/components/sections/date-time-picker/time-picker-ng1/wrapper/time-picker-wrapper.directive.html
@@ -1,3 +1,7 @@
 <timepicker ng-model="vm.mytime" ng-change="vm.changed()" hour-step="vm.hstep" minute-step="vm.mstep" 
   show-meridian="vm.ismeridian">
 </timepicker>
+
+<p class="text-error" ng-if="vm.mytime === null">
+  <span class="hpe-icon hpe-alert"></span> Enter a valid time.
+</p>
