+++ docs/app/pages/components/sections/input-controls/number-picker/number-picker.component.html
@@ -1,8 +1,18 @@
 <p>Selecting an integer:</p>
-<ux-number-picker class="number-picker-component-demo" min="-10" max="10"></ux-number-picker>
+
+<ux-number-picker class="number-picker-component-demo" min="-10" max="10" [valid]="form.controls['integer'].valid" [formControl]="form.controls['integer']"></ux-number-picker>
+
+<p class="text-error" *ngIf="form.controls['integer'].invalid">
+    <span class="hpe-icon hpe-alert"></span> Enter a number between -10 and 10.
+</p>
 
 <p class="m-t-md">Selecting a decimal:</p>
-<ux-number-picker class="number-picker-component-demo" step="0.5" min="0" max="10"></ux-number-picker>
+
+<ux-number-picker class="number-picker-component-demo" step="0.5" min="0" max="10" [valid]="form.controls['decimal'].valid" [formControl]="form.controls['decimal']"></ux-number-picker>
+
+<p class="text-error" *ngIf="form.controls['decimal'].invalid">
+    <span class="hpe-icon hpe-alert"></span> Enter a number between 0 and 10.
+</p>
 
 <hr>
 
@@ -56,15 +66,15 @@
             <td><b>@Input</b> valid</td>
             <td><b>boolean</b></td>
             <td>
-                Can be used to show a red outline around the input to indicate an invalid value. By default the error state will
-                appear if the user enters a number below the minimum value or above the maximum value.
+                Can be used to show a red outline around the input to indicate an invalid value. By default the error state will appear if
+                the user enters a number below the minimum value or above the maximum value.
             </td>
         </tr>
 
     </tbody>
 </table>
 
-<p>The number picker component can be disabled using the <code>disabled</code> attribute. It also supports <code>ngModel</code> for use within forms.</p>
+<p>The number picker component can be disabled using the <code>disabled</code> attribute. It also supports <code>ngModel</code>    for use within forms.</p>
 
 <p>The following code can be used to create the example shown above:</p>
 
@@ -72,7 +82,10 @@
     <tab heading="HTML">
         <uxd-snippet [content]="snippets.compiled.appHtml"></uxd-snippet>
     </tab>
+    <tab heading="TypeScript">
+        <uxd-snippet [content]="snippets.compiled.appTs"></uxd-snippet>
+    </tab>
     <tab heading="CSS">
-        <uxd-snippet [content]="snippets.compiled.appCss"></uxd-snippet>        
+        <uxd-snippet [content]="snippets.compiled.appCss"></uxd-snippet>
     </tab>
 </tabset>
