+++ docs/app/pages/components/sections/input-controls/number-picker/number-picker.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { FormBuilder, Validators, FormGroup } from '@angular/forms';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
 import { BaseDocumentationSection } from '../../../../../components/base-documentation-section/base-documentation-section';
 import { IPlunkProvider } from '../../../../../interfaces/IPlunkProvider';
@@ -10,10 +11,12 @@ import { IPlunk } from '../../../../../interfaces/IPlunk';
 })
 @DocumentationSectionComponent('ComponentsNumberPickerComponent')
 export class ComponentsNumberPickerComponent extends BaseDocumentationSection implements IPlunkProvider {
+    
+    form: FormGroup;
 
     plunk: IPlunk = {
         files: {
-            'app.component.ts': this.snippets.raw.appExampleTs,
+            'app.component.ts': this.snippets.raw.appTs,
             'app.component.html': this.snippets.raw.appHtml,
             'app.component.css': this.snippets.raw.appCss
         },
@@ -23,7 +26,12 @@ export class ComponentsNumberPickerComponent extends BaseDocumentationSection im
         }]
     };
 
-    constructor() {
+    constructor(formBuilder: FormBuilder) {
         super(require.context('./snippets/', false, /\.(html|css|js|ts)$/));
+
+        this.form = formBuilder.group({
+            integer: [0, Validators.compose([Validators.required, Validators.min(-10), Validators.max(10)])],
+            decimal: [0, Validators.compose([Validators.required, Validators.min(0), Validators.max(10)])]
+        });
     }
 }
