--- docs/app/pages/components/sections/input-controls/number-picker/snippets/app.example.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'app',
-    templateUrl: './src/app.component.html',
-    styleUrls: ['./src/app.component.css']
-})
-export class AppComponent {
-
-}
