+++ docs/app/pages/components/sections/input-controls/tags/snippets/app.html
@@ -26,10 +26,12 @@
 
             </ux-tag-input>
             <span *ngIf="tagInput.validationErrors.tagRangeError" class="validation-error-text">
+                <span class="hpe-icon hpe-alert"></span> 
                 Required: number of tags between {{tagInput.validationErrors.tagRangeError.min}}
                 and {{tagInput.validationErrors.tagRangeError.max}}.
             </span>
             <span *ngIf="tagInput.validationErrors.inputPattern" class="validation-error-text">
+                <span class="hpe-icon hpe-alert"></span> 
                 Expected format: {{tagInput.validationErrors.inputPattern.pattern.source}}
             </span>
         </form>
