+++ docs/app/services/plunker/templates/index_html.txt
@@ -15,7 +15,7 @@
         window.uxdAssetsUrl = '${assetsUrl}';
     </script>
     <script src="https://unpkg.com/es6-shim@0.35.3"></script>
-    <script src="https://unpkg.com/zone.js/dist/zone.js"></script>
+    <script src="https://unpkg.com/zone.js@0.8.12/dist/zone.js"></script>
     <script src="https://unpkg.com/zone.js/dist/long-stack-trace-zone.js"></script>
     <script src="https://unpkg.com/reflect-metadata@0.1.3/Reflect.js"></script>
     <script src="https://unpkg.com/systemjs@0.19.31/dist/system.js"></script>
