+++ docs/app/services/plunker/templates/main_ts.txt
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppComponent } from './app.component';
 ${imports}
@@ -11,7 +11,6 @@ ${imports}
     bootstrap: [AppComponent]
 })
 export class AppModule {
-    ngDoBootstrap() { }
 }
 
 platformBrowserDynamic().bootstrapModule(AppModule);
